chore: restructure @waku/sdk
