+++ packages/sdk/src/index.ts
@@ -10,7 +10,7 @@ export { utf8ToBytes, bytesToUtf8 } from "@waku/utils/bytes";
 export { defaultLibp2p } from "./utils/libp2p.js";
 export * from "./utils/content_topic.js";
 export * from "./waku.js";
-export * from "./create.js";
+export * from "./node/light.js";
 export * as waku from "@waku/core";
 export * as utils from "@waku/utils";
 export * from "@waku/interfaces";
